move dataclass to skylet and use dataclass inside command runner
