+++ sky/resources.py
@@ -11,6 +11,7 @@ from sky import skypilot_config
 from sky import spot
 from sky.backends import backend_utils
 from sky.skylet import constants
+from sky.skylet.providers import command_runner
 from sky.utils import accelerator_registry
 from sky.utils import schemas
 from sky.utils import tpu_utils
@@ -53,8 +54,7 @@ class Resources:
         disk_tier: Optional[Literal['high', 'medium', 'low']] = None,
         ports: Optional[List[Union[int, str]]] = None,
         # Internal use only.
-        _docker_login_config: Optional[
-            'backend_utils.DockerLoginConfig'] = None,
+        _docker_login_config: Optional[command_runner.DockerLoginConfig] = None,
         _is_image_managed: Optional[bool] = None,
     ):
         """Initialize a Resources object.
