+++ sky/skylet/providers/azure/node_provider.py
@@ -16,6 +16,7 @@ from sky.skylet.providers.azure.config import (
     get_azure_sdk_function,
 )
 from sky.skylet.providers.command_runner import SkyDockerCommandRunner
+from sky.skylet.providers.command_runner import DockerLoginConfig
 
 from ray.autoscaler._private.command_runner import SSHCommandRunner
 from ray.autoscaler.node_provider import NodeProvider
@@ -85,9 +86,6 @@ class AzureNodeProvider(NodeProvider):
 
         # cache node objects
         self.cached_nodes = {}
-
-        # Set docker login config if it exists
-        self.docker_login_config = provider_config.get("docker_login_config", None)
 
     @synchronized
     def _get_filtered_nodes(self, tag_filters):
@@ -446,8 +444,10 @@ class AzureNodeProvider(NodeProvider):
             "use_internal_ip": use_internal_ip,
         }
         if docker_config and docker_config["container_name"] != "":
-            if self.docker_login_config is not None:
-                docker_config.update(self.docker_login_config)
+            if "docker_login_config" in self.provider_config:
+                docker_config["docker_login_config"] = DockerLoginConfig.from_dict(
+                    self.provider_config["docker_login_config"]
+                )
             return SkyDockerCommandRunner(docker_config, **common_args)
         else:
             return SSHCommandRunner(**common_args)
