+++ sky/skylet/providers/gcp/node_provider.py
@@ -13,6 +13,7 @@ from sky.skylet.providers.gcp.config import (
     get_node_type,
 )
 from sky.skylet.providers.command_runner import SkyDockerCommandRunner
+from sky.skylet.providers.command_runner import DockerLoginConfig
 
 from ray.autoscaler._private.command_runner import SSHCommandRunner
 from ray.autoscaler.tags import (
@@ -81,9 +82,6 @@ class GCPNodeProvider(NodeProvider):
         # excessive DescribeInstances requests.
         self.cached_nodes: Dict[str, GCPNode] = {}
         self.cache_stopped_nodes = provider_config.get("cache_stopped_nodes", True)
-
-        # Set docker login config if it exists
-        self.docker_login_config = provider_config.get("docker_login_config", None)
 
     def _construct_clients(self):
         _, _, compute, tpu = construct_clients_from_provider_config(
@@ -390,8 +388,10 @@ class GCPNodeProvider(NodeProvider):
             "use_internal_ip": use_internal_ip,
         }
         if docker_config and docker_config["container_name"] != "":
-            if self.docker_login_config is not None:
-                docker_config.update(self.docker_login_config)
+            if "docker_login_config" in self.provider_config:
+                docker_config["docker_login_config"] = DockerLoginConfig.from_dict(
+                    self.provider_config["docker_login_config"]
+                )
             return SkyDockerCommandRunner(docker_config, **common_args)
         else:
             return SSHCommandRunner(**common_args)
